TOT-96 : [REFACTOR] 계획 작성 요청 Dto 클래스들을 하나의 Dto로 수정.
- PlanRequestDto 클래스 내부에 PlanCreateRequestDto, ScheduleCreateRequestDto, SchedulePlaceCreateRequestDto 내부 클래스들을 추가.
- 기존 Dto들 삭제.
